+++ smartcontract/cli/src/lib.rs
@@ -31,6 +31,7 @@ pub mod subscribe;
 pub mod tenant;
 pub mod tests;
 pub mod topology;
+pub mod transfer;
 pub mod user;
 pub mod util;
 pub mod validators;